fix: rename FdmSettings to OctoPrintSettings
